+++ package.json
@@ -5,12 +5,13 @@
   "version": "0.1.0",
   "author": "Daphne Charles <daphneacharles@gmail.com>",
   "dependencies": {
+    "babel-plugin-styled-components": "1.10.6",
     "gatsby": "^2.7.1",
     "gatsby-image": "^2.1.0",
     "gatsby-plugin-manifest": "^2.1.1",
-    "gatsby-plugin-offline": "^2.1.1",
     "gatsby-plugin-react-helmet": "^3.0.12",
     "gatsby-plugin-sharp": "^2.1.0",
+    "gatsby-plugin-styled-components": "3.1.3",
     "gatsby-plugin-typography": "2.2.13",
     "gatsby-remark-images": "3.1.0",
     "gatsby-remark-relative-images": "0.2.2",
