+++ src/pages/404.js
@@ -6,7 +6,7 @@ import SEO from "../components/seo"
 
 const NotFoundPage = () => (
   <Layout>
-    <SEO title="404: Not found" />
+    <SEO title="Page Not  | Daphne Charles" />
     <h1>PAGE NOT FOUND</h1>
     <p>You just navigated to a page that doesn't exist. Oopsies!</p>
     <Link to="/">Click here to head home.</Link>
